+++ src/store/user-data.ts
@@ -76,6 +76,7 @@ export const initialUserData = deepFreeze({
     order: [] as string[],
     hidden: [] as string[],
     locked: [] as string[],
+    folders: [] as UserData.TabFolder[],
   },
   commandLists: [] as UserData.CommandList[],