+++ src/store/user-data.types.d.ts
@@ -136,4 +136,10 @@ declare namespace UserData {
   }
 
   type ExchangeChartType = 'SMOOTH' | 'ALIGNED' | 'RAW';
+
+  interface TabFolder {
+    id: string;
+    name: string;
+    screenIds: string[];
+  }
 }